Add analyze command
The idea behind this command is to avoid generating all rules analyzer
rules every single time we run any dbt command.

In the future, caching the generator output is probably even nicer, but
this is a lower hanging fruit that could easily be fixed to avoid paying
the cost of sa-report targets if we are not attempting to use them.

Note that the second pass has not yet changed to be only run in coverage
and analyze modes because it would be a backwards incompatible change.

Once all jobs are updated to use the new analyze command we can add that
without breaking the jobs.