Fail build if multiple dbt modules provide the same BUILD.go file
This can now happen because the dbt module and the Go module names are
no longer tied in the CPP layout.

So far, when this has happened, one of the 2 files would be taken
randomly (Because modules are part of a hash map which has random
ordering in Go) and has non-repeatable behavior.

This makes it consistent so that as soon as a collision is detected, the
operation is aborted and the user is notified.