Fix for the legacy layout
The cpp layout decouples the dbt Module name and the GoModule names such
that they don't need to be shared. However, this is not true in the legacy
layout and therefore changing the name of the git module (as was done in
https://gitea.daedalean.ai/daedalean-github/dbt/pulls/8) results in a
broken workflow if the name of the workspace folder does not match the
name of the repository.

In this change, I revert the change to the GitModule name (which made
it behave the same as the TarModule) and instead enforce that the names
of dependencies listed in a MODULE file are the same as that of the
repository (which was the original motivation for the change I did for
the manifest support).

In practice it is a shame that the workflow used by the FPGA and SW
teams is diverging to this degree, but should fix the issue for now